// src/resources/matrix/rooms/events.ts
export class BaseEvents extends APIResource {
  static override readonly _key: readonly ['matrix', 'rooms', 'events'] = Object.freeze([
    'matrix',
    'rooms',
    'events',
  ] as const);

  /**
   * Get a single event based on `roomId/eventId`. You must have permission to
   * retrieve this event e.g. by being a member in the room for this event.
   *
   * @example
   * ```ts
   * const event = await client.matrix.rooms.events.retrieve(
   *   '$asfDuShaf7Gafaw:matrix.org',
   *   { roomId: '!636q39766251:matrix.org' },
   * );
   * ```
   */
  retrieve(
    eventID: string,
    params: EventRetrieveParams,
    options?: RequestOptions,
  ): APIPromise<EventRetrieveResponse> {
    const { roomId } = params;
    return this._client.get(path`/_matrix/client/v3/rooms/${roomId}/event/${eventID}`, options);
  }
}
export class Events extends BaseEvents {}

/**
 * The format used for events when they are returned from a homeserver to a client
 * via the Client-Server API, or sent to an Application Service via the Application
 * Services API.
 */
export interface EventRetrieveResponse {
  /**
   * The body of this event, as created by the client which sent it.
   */
  content: unknown;

  /**
   * The globally unique identifier for this event.
   */
  event_id: string;

  /**
   * Timestamp (in milliseconds since the unix epoch) on originating homeserver when
   * this event was sent.
   */
  origin_server_ts: number;

  /**
   * The ID of the room associated with this event.
   */
  room_id: string;

  /**
   * Contains the fully-qualified ID of the user who sent this event.
   */
  sender: string;

  /**
   * The type of the event.
   */
  type: string;

  /**
   * Present if, and only if, this event is a _state_ event. The key making this
   * piece of state unique in the room. Note that it is often an empty string.
   *
   * State keys starting with an `@` are reserved for referencing user IDs, such as
   * room members. With the exception of a few events, state events set with a given
   * user's ID as the state key MUST only be set by that user.
   */
  state_key?: string;

  unsigned?: EventRetrieveResponse.Unsigned;
}

export namespace EventRetrieveResponse {
  export interface Unsigned {
    /**
     * The time in milliseconds that has elapsed since the event was sent. This field
     * is generated by the local homeserver, and may be incorrect if the local time on
     * at least one of the two servers is out of sync, which can cause the age to
     * either be negative or greater than it actually is.
     */
    age?: number;

    /**
     * The room membership of the user making the request, at the time of the event.
     *
     * This property is the value of the `membership` property of the requesting user's
     * [`m.room.member`](https://spec.matrix.org/v1.18/client-server-api#mroommember)
     * state at the point of the event, including any changes caused by the event. If
     * the user had yet to join the room at the time of the event (i.e, they have no
     * `m.room.member` state), this property is set to `leave`.
     *
     * Homeservers SHOULD populate this property wherever practical, but they MAY omit
     * it if necessary (for example, if calculating the value is expensive, servers
     * might choose to only implement it in encrypted rooms). The property is _not_
     * normally populated in events pushed to application services via the application
     * service transaction API (where there is no clear definition of "requesting
     * user").
     */
    membership?: string;

    /**
     * The previous `content` for this event. This field is generated by the local
     * homeserver, and is only returned if the event is a state event, and the client
     * has permission to see the previous content.
     */
    prev_content?: unknown;

    redacted_because?: unknown;

    /**
     * The client-supplied
     * [transaction ID](https://spec.matrix.org/v1.18/client-server-api/#transaction-identifiers),
     * for example, provided via
     * `PUT /_matrix/client/v3/rooms/{roomId}/send/{eventType}/{txnId}`, if the client
     * being given the event is the same one which sent it.
     */
    transaction_id?: string;
  }
}

export interface EventRetrieveParams {
  /**
   * The ID of the room the event is in.
   */
  roomId: string;
}

export declare namespace Events {
  export {
    type EventRetrieveResponse as EventRetrieveResponse,
    type EventRetrieveParams as EventRetrieveParams,
  };
}