// src/resources/matrix/rooms/state.ts
export class BaseState extends APIResource {
  static override readonly _key: readonly ['matrix', 'rooms', 'state'] = Object.freeze([
    'matrix',
    'rooms',
    'state',
  ] as const);

  /**
   * Looks up the contents of a state event in a room. If the user is joined to the
   * room then the state is taken from the current state of the room. If the user has
   * left the room then the state is taken from the state of the room when they left.
   *
   * @example
   * ```ts
   * const state = await client.matrix.rooms.state.retrieve(
   *   'state_key',
   *   {
   *     roomId: '!636q39766251:example.com',
   *     eventType: 'm.room.name',
   *   },
   * );
   * ```
   */
  retrieve(
    stateKey: string,
    params: StateRetrieveParams,
    options?: RequestOptions,
  ): APIPromise<StateRetrieveResponse> {
    const { roomId, eventType, ...query } = params;
    return this._client.get(path`/_matrix/client/v3/rooms/${roomId}/state/${eventType}/${stateKey}`, {
      query,
      ...options,
    });
  }

  /**
   * Get the state events for the current state of a room.
   *
   * @example
   * ```ts
   * const states = await client.matrix.rooms.state.list(
   *   '!636q39766251:example.com',
   * );
   * ```
   */
  list(roomID: string, options?: RequestOptions): APIPromise<StateListResponse> {
    return this._client.get(path`/_matrix/client/v3/rooms/${roomID}/state`, options);
  }
}
export class State extends BaseState {}

export type StateRetrieveResponse = { [key: string]: unknown };

/**
 * If the user is a member of the room this will be the current state of the room
 * as a list of events. If the user has left the room then this will be the state
 * of the room when they left as a list of events.
 */
export type StateListResponse = Array<StateListResponse.StateListResponseItem>;

export namespace StateListResponse {
  /**
   * The format used for events when they are returned from a homeserver to a client
   * via the Client-Server API, or sent to an Application Service via the Application
   * Services API.
   */
  export interface StateListResponseItem {
    /**
     * The body of this event, as created by the client which sent it.
     */
    content: unknown;

    /**
     * The globally unique identifier for this event.
     */
    event_id: string;

    /**
     * Timestamp (in milliseconds since the unix epoch) on originating homeserver when
     * this event was sent.
     */
    origin_server_ts: number;

    /**
     * The ID of the room associated with this event.
     */
    room_id: string;

    /**
     * Contains the fully-qualified ID of the user who sent this event.
     */
    sender: string;

    /**
     * The type of the event.
     */
    type: string;

    /**
     * Present if, and only if, this event is a _state_ event. The key making this
     * piece of state unique in the room. Note that it is often an empty string.
     *
     * State keys starting with an `@` are reserved for referencing user IDs, such as
     * room members. With the exception of a few events, state events set with a given
     * user's ID as the state key MUST only be set by that user.
     */
    state_key?: string;

    unsigned?: StateListResponseItem.Unsigned;
  }

  export namespace StateListResponseItem {
    export interface Unsigned {
      /**
       * The time in milliseconds that has elapsed since the event was sent. This field
       * is generated by the local homeserver, and may be incorrect if the local time on
       * at least one of the two servers is out of sync, which can cause the age to
       * either be negative or greater than it actually is.
       */
      age?: number;

      /**
       * The room membership of the user making the request, at the time of the event.
       *
       * This property is the value of the `membership` property of the requesting user's
       * [`m.room.member`](https://spec.matrix.org/v1.18/client-server-api#mroommember)
       * state at the point of the event, including any changes caused by the event. If
       * the user had yet to join the room at the time of the event (i.e, they have no
       * `m.room.member` state), this property is set to `leave`.
       *
       * Homeservers SHOULD populate this property wherever practical, but they MAY omit
       * it if necessary (for example, if calculating the value is expensive, servers
       * might choose to only implement it in encrypted rooms). The property is _not_
       * normally populated in events pushed to application services via the application
       * service transaction API (where there is no clear definition of "requesting
       * user").
       */
      membership?: string;

      /**
       * The previous `content` for this event. This field is generated by the local
       * homeserver, and is only returned if the event is a state event, and the client
       * has permission to see the previous content.
       */
      prev_content?: unknown;

      redacted_because?: unknown;

      /**
       * The client-supplied
       * [transaction ID](https://spec.matrix.org/v1.18/client-server-api/#transaction-identifiers),
       * for example, provided via
       * `PUT /_matrix/client/v3/rooms/{roomId}/send/{eventType}/{txnId}`, if the client
       * being given the event is the same one which sent it.
       */
      transaction_id?: string;
    }
  }
}

export interface StateRetrieveParams {
  /**
   * Path param: The room to look up the state in.
   */
  roomId: string;

  /**
   * Path param: The type of state to look up.
   */
  eventType: string;

  /**
   * Query param: The format to use for the returned data. `content` (the default)
   * will return only the content of the state event. `event` will return the entire
   * event in the usual format suitable for clients, including fields like event ID,
   * sender and timestamp.
   */
  format?: 'content' | 'event';
}

export declare namespace State {
  export {
    type StateRetrieveResponse as StateRetrieveResponse,
    type StateListResponse as StateListResponse,
    type StateRetrieveParams as StateRetrieveParams,
  };
}